wip: refactor to create `MagicMove`, add `ProcessPulses` steps
