+++ package.json
@@ -49,12 +49,14 @@
 		"remark-gfm": "^4.0.0",
 		"remark-toc": "^9.0.0",
 		"shiki-magic-move": "^0.4.3",
+		"type-fest": "^4.23.0",
 		"unist-util-visit": "^5.0.0",
 		"zustand": "^4.5.4"
 	},
 	"devDependencies": {
 		"@notionhq/client": "^2.2.15",
 		"@octokit/types": "^13.5.0",
+		"@types/eslint": "~9.6.0",
 		"@types/lodash-es": "^4.17.12",
 		"@types/mdx": "^2.0.13",
 		"@types/node": "^20.14.12",
